+++ lib/l10n/app_localizations_en.dart
@@ -245,6 +245,12 @@ class AppLocalizationsEn extends AppLocalizations {
   String get helpSupportCopySupportIdTooltip => 'Copy support ID';
 
   @override
+  String get helpSupportRoadmapLabel => 'Roadmap';
+
+  @override
+  String get helpSupportRoadmapValue => 'View what’s coming';
+
+  @override
   String helpSupportAppVersionRow(Object version) {
     return 'App version $version';
   }
@@ -324,6 +330,12 @@ class AppLocalizationsEn extends AppLocalizations {
 
   @override
   String get helpSupportXTwitterLabel => 'X / Twitter';
+
+  @override
+  String get helpSupportInstagramLabel => 'Instagram';
+
+  @override
+  String get helpSupportMastodonLabel => 'Mastodon';
 
   @override
   String get helpSupportThreadsLabel => 'Threads';